fix & style: updating splatalogue query and passing lint
Implemented a dummy `correlate_experiments` function just to
get the static analysis to pass. It wasn't working before,
so I don't think it'll need to work immediately.

Splatalogue has also completely broken astroquery...again.
See https://github.com/astropy/astroquery/issues/2956 for
the discussion. In the meantime, the astroquery will also
need the main branch install instead of a release for this
to work: the headers are parsed from Splatalogue, but the
data is completely garbage.