fix: updating html output based on pandas API changes
The `.format` function now returns a `StyleRender` object, which
is different to what it used to be, removing the ability to render the
output as HTML (at least obviously so).

For now, this commit will ensure functionality in exchange for a lot
more digits of precision than needed in the HTML output.